noise gen

+++ gamedev-utils/src/utils/NoiseGenerator.java
@@ -32,7 +32,7 @@ public class NoiseGenerator {
 	 *            the second octave for perlin
 	 * @return a 2D array of perlin noise
 	 */
-	public double[][] getPerlinNoise(int width, int height, int octave1, int octave2) {
+	public float[][] getPerlinNoise(int width, int height, int octave1, int octave2) {
 		this.width = width;
 		this.height = height;
 		return genPerlinNoise(genSmoothNoisePerlin(genWhiteNoise(width, height), octave1), octave2);
@@ -50,7 +50,7 @@ public class NoiseGenerator {
 	 *            the number of times the smoothing process is repeated
 	 * @return a 2D array of noise
 	 */
-	public double[][] getSmoothNoise(int width, int height, int octave) {
+	public float[][] getSmoothNoise(int width, int height, int octave) {
 		this.width = width;
 		this.height = height;
 		return genSmoothNoise(genWhiteNoise(width, height), octave);
@@ -79,49 +79,49 @@ public class NoiseGenerator {
 		return genCellularAutomataNoise(width, height, birthLimit, deathLimit, numberOfSteps);
 	}
 
-	private double[][] genWhiteNoise(int width, int height) {
-		double[][] noise = new double[width][height];
+	private float[][] genWhiteNoise(int width, int height) {
+		float[][] noise = new float[width][height];
 		for (int i = 0; i < width; i++) {
 			for (int j = 0; j < height; j++) {
-				noise[i][j] = random.nextDouble();
+				noise[i][j] = random.nextFloat();
 			}
 		}
 		return noise;
 	}
 
-	private double[][] genSmoothNoise(double[][] whiteNoise, int octaves) {
-		double[][] smoothNoise = new double[width][height];
+	private float[][] genSmoothNoise(float[][] whiteNoise, int octaves) {
+		float[][] smoothNoise = new float[width][height];
 		for (int repeat = 0; repeat < octaves; repeat++) {
-			double[][] noise = new double[width][height];
+			float[][] noise = new float[width][height];
 			for (int i = 0; i < width - 1; i++) {
 				for (int j = 0; j < height - 1; j++) {
 					if (i > 0 && j > 0 && i < width && j < height) {
-						noise[i][j] = MathHelper.smooth9(whiteNoise[i - 1][j - 1], whiteNoise[i - 1][j],
+						noise[i][j] = (float) MathHelper.smooth9(whiteNoise[i - 1][j - 1], whiteNoise[i - 1][j],
 								whiteNoise[i - 1][j + 1], whiteNoise[i][j - 1], whiteNoise[i][j], whiteNoise[i][j + 1],
 								whiteNoise[i + 1][j - 1], whiteNoise[i + 1][j], whiteNoise[i + 1][j + 1]);
 					} else if (i == 0 && j == 0) {
-						noise[i][j] = MathHelper.smooth4(whiteNoise[i][j], whiteNoise[i][j + 1], whiteNoise[i + 1][j],
+						noise[i][j] = (float) MathHelper.smooth4(whiteNoise[i][j], whiteNoise[i][j + 1], whiteNoise[i + 1][j],
 								whiteNoise[i + 1][j + 1]);
 					} else if (i == 0 && j == height) {
-						noise[i][j] = MathHelper.smooth4(whiteNoise[i][j - 1], whiteNoise[i][j],
+						noise[i][j] = (float) MathHelper.smooth4(whiteNoise[i][j - 1], whiteNoise[i][j],
 								whiteNoise[i + 1][j - 1], whiteNoise[i + 1][j]);
 					} else if (i == width && j == 0) {
-						noise[i][j] = MathHelper.smooth4(whiteNoise[i - 1][j], whiteNoise[i - 1][j + 1],
+						noise[i][j] = (float) MathHelper.smooth4(whiteNoise[i - 1][j], whiteNoise[i - 1][j + 1],
 								whiteNoise[i][j], whiteNoise[i][j + 1]);
 					} else if (i == 0) {
-						noise[i][j] = MathHelper.smooth6(whiteNoise[i][j - 1], whiteNoise[i][j], whiteNoise[i][j + 1],
+						noise[i][j] = (float) MathHelper.smooth6(whiteNoise[i][j - 1], whiteNoise[i][j], whiteNoise[i][j + 1],
 								whiteNoise[i + 1][j - 1], whiteNoise[i + 1][j], whiteNoise[i + 1][j + 1]);
 					} else if (j == 0) {
-						noise[i][j] = MathHelper.smooth6(whiteNoise[i - 1][j], whiteNoise[i - 1][j + 1],
+						noise[i][j] = (float) MathHelper.smooth6(whiteNoise[i - 1][j], whiteNoise[i - 1][j + 1],
 								whiteNoise[i][j], whiteNoise[i][j + 1], whiteNoise[i + 1][j], whiteNoise[i + 1][j + 1]);
 					} else if (i == width && j == height) {
-						noise[i][j] = MathHelper.smooth4(whiteNoise[i - 1][j - 1], whiteNoise[i - 1][j],
+						noise[i][j] = (float) MathHelper.smooth4(whiteNoise[i - 1][j - 1], whiteNoise[i - 1][j],
 								whiteNoise[i][j - 1], whiteNoise[i][j]);
 					} else if (i == width) {
-						noise[i][j] = MathHelper.smooth6(whiteNoise[i - 1][j - 1], whiteNoise[i - 1][j],
+						noise[i][j] = (float) MathHelper.smooth6(whiteNoise[i - 1][j - 1], whiteNoise[i - 1][j],
 								whiteNoise[i - 1][j + 1], whiteNoise[i][j - 1], whiteNoise[i][j], whiteNoise[i][j + 1]);
 					} else if (j == height) {
-						noise[i][j] = MathHelper.smooth6(whiteNoise[i - 1][j - 1], whiteNoise[i - 1][j],
+						noise[i][j] = (float) MathHelper.smooth6(whiteNoise[i - 1][j - 1], whiteNoise[i - 1][j],
 								whiteNoise[i][j - 1], whiteNoise[i][j], whiteNoise[i + 1][j - 1], whiteNoise[i + 1][j]);
 					}
 					smoothNoise[i][j] += noise[i][j];
@@ -136,37 +136,37 @@ public class NoiseGenerator {
 		return smoothNoise;
 	}
 
-	private double[][] genSmoothNoisePerlin(double[][] baseNoise, int octave) {
-		double[][] smoothNoise = new double[width][height];
+	private float[][] genSmoothNoisePerlin(float[][] baseNoise, int octave) {
+		float[][] smoothNoise = new float[width][height];
 		int samplePeriod = 1 << octave;
-		double sampleFrequency = 1.0f / samplePeriod;
+		float sampleFrequency = 1.0f / samplePeriod;
 		for (int i = 0; i < width; i++) {
 			int sample_i0 = (i / samplePeriod) * samplePeriod;
 			int sample_i1 = (sample_i0 + samplePeriod) % width;
-			double horizontal_blend = (i - sample_i0) * sampleFrequency;
+			float horizontal_blend = (i - sample_i0) * sampleFrequency;
 			for (int j = 0; j < height; j++) {
 				int sample_j0 = (j / samplePeriod) * samplePeriod;
 				int sample_j1 = (sample_j0 + samplePeriod) % height;
-				double vertical_blend = (j - sample_j0) * sampleFrequency;
-				double top = MathHelper.cosineInterpolate(baseNoise[sample_i0][sample_j0],
+				float vertical_blend = (j - sample_j0) * sampleFrequency;
+				float top = (float) MathHelper.cosineInterpolate(baseNoise[sample_i0][sample_j0],
 						baseNoise[sample_i1][sample_j0], horizontal_blend);
-				double bottom = MathHelper.cosineInterpolate(baseNoise[sample_i0][sample_j1],
+				float bottom = (float) MathHelper.cosineInterpolate(baseNoise[sample_i0][sample_j1],
 						baseNoise[sample_i1][sample_j1], horizontal_blend);
-				smoothNoise[i][j] = MathHelper.cosineInterpolate(top, bottom, vertical_blend);
+				smoothNoise[i][j] = (float) MathHelper.cosineInterpolate(top, bottom, vertical_blend);
 			}
 		}
 		return smoothNoise;
 	}
 
-	private double[][] genPerlinNoise(double[][] baseNoise, int octaveCount) {
-		double[][][] smoothNoise = new double[octaveCount][][];
-		double persistance = 0.5f;
+	private float[][] genPerlinNoise(float[][] baseNoise, int octaveCount) {
+		float[][][] smoothNoise = new float[octaveCount][][];
+		float persistance = 0.5f;
 		for (int i = 0; i < octaveCount; i++) {
 			smoothNoise[i] = genSmoothNoisePerlin(baseNoise, i);
 		}
-		double[][] perlinNoise = new double[width][height];
-		double amplitude = 1.0f;
-		double totalAmplitude = 0.0f;
+		float[][] perlinNoise = new float[width][height];
+		float amplitude = 1.0f;
+		float totalAmplitude = 0.0f;
 		for (int octave = octaveCount - 1; octave >= 0; octave--) {
 			amplitude *= persistance;
 			totalAmplitude += amplitude;
@@ -261,29 +261,32 @@ public class NoiseGenerator {
 		return s.toString();
 	}
 
-	public Rectangle[] generateRooms(int numOfRooms, int xLimit, int yLimit, int xMod, int yMod, int widthLimit, int heightLimit, int widthMod, int heightMod, int corridorSize) {
+	public Rectangle[] generateRooms(int numOfRooms, int xLimit, int yLimit, int xMod, int yMod, int widthLimit,
+			int heightLimit, int widthMod, int heightMod, int corridorSize) {
 		ArrayList<Rectangle> rooms = new ArrayList<Rectangle>();
-		for(int i = 0; i < numOfRooms; i++){
-			rooms.add(new Rectangle(MathHelper.random.nextInt(xLimit) + xMod, MathHelper.random.nextInt(yLimit) + yMod, MathHelper.random.nextInt(widthLimit) + widthMod, MathHelper.random.nextInt(heightLimit) + heightMod));
+		for (int i = 0; i < numOfRooms; i++) {
+			rooms.add(new Rectangle(MathHelper.random.nextInt(xLimit) + xMod, MathHelper.random.nextInt(yLimit) + yMod,
+					MathHelper.random.nextInt(widthLimit) + widthMod,
+					MathHelper.random.nextInt(heightLimit) + heightMod));
 		}
 		int pls = rooms.size();
-		for(int i = 0; i < pls; i++){
-			for(int j = 0; j < pls; j++){
+		for (int i = 0; i < pls; i++) {
+			for (int j = 0; j < pls; j++) {
 				Rectangle r1 = rooms.get(i), r2 = rooms.get(j);
 				int x, y, width, height;
-				if(r1.x < r2.x){
-					x = r1.x + r1.width/2;
-					width = (r2.x + r2.width/2) - x;
-				}else{
-					x = r2.x + r2.width/2;
-					width = (r1.x + r1.width/2) - x;
+				if (r1.x < r2.x) {
+					x = r1.x + r1.width / 2;
+					width = (r2.x + r2.width / 2) - x;
+				} else {
+					x = r2.x + r2.width / 2;
+					width = (r1.x + r1.width / 2) - x;
 				}
-				if(r1.y < r2.y){
-					y = r1.y + r1.height/2;
-					height = (r2.y + r2.height/2) - y;
-				}else{
-					y = r2.y + r2.height/2;
-					height = (r1.y + r1.height/2) - y;
+				if (r1.y < r2.y) {
+					y = r1.y + r1.height / 2;
+					height = (r2.y + r2.height / 2) - y;
+				} else {
+					y = r2.y + r2.height / 2;
+					height = (r1.y + r1.height / 2) - y;
 				}
 				rooms.add(new Rectangle(x, y, width, corridorSize));
 				rooms.add(new Rectangle(x, y, corridorSize, height));
